refactor(MenuView): 메뉴View클래스들을 계층화
메뉴View들이 모두 엇 비슷한 기능을 하는데도, 관리 기능 View 만이 추상클래스 ManagementView에 의해 하위 클래스로 묶여있고 MainView는 독립적으로 존재했습니다. 이는 유저에게서 메뉴를 선택하도록 요구하고, 그 값에 따라 리턴하는 타입이 관리 뷰는 FunctionType이고 메인 화면은 MainMenuType로 서로 다르다는 점 때문에 존재하는 난해함 때문이었습니다.
따라서 제네릭 추상 클래스 MenuView를 만들었고, 이를 상속할 때 어떤 메뉴 타입(현재까지 FunctionType, MainMenuType 2개 존재)을 리턴할 것인지 명시하도록 했습니다. 이에 메뉴 뷰들의 상속 구조는 다음과 같습니다.

MenuView(Abstract) -> MainView
                   -> ManagementView(Abstract) -> Station/Line/SectionView

즉, 메인 화면은 바로 MenuView를 상속받고, 관리 기능 화면들은 MenuView를 상속받은 ManagementView를 상속하여 만듭니다.
결과적으로 계층화를 통하여 공통 로직을 묶었습니다.
