+++ src/main/java/subway/view/menuView/StationView.java
@@ -1,8 +1,10 @@
-package subway.view.managementView;
+package subway.view.menuView;
 
 import subway.menuType.FunctionType;
 
 import java.util.*;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
 
 public class StationView extends ManagementView {
     private static final String ITEM_PREFIX = "역 ";
@@ -15,17 +17,24 @@ public class StationView extends ManagementView {
     private static StationView instance;
 
     private StationView() {
-        List<String> descriptions = new ArrayList<>(Arrays.asList(
-                MENU_CREATE, MENU_DELETE, MENU_READ
-        ));
-        initializeSelections(descriptions);
-
-        initializeHashMapToFunctionType(selections.toList(), Arrays.asList(FunctionType.values()));
+        initializeSelections();
+        initializeHashMapToFunctionType(this.selections.toList(), Arrays.asList(FunctionType.values()));
 
         viewName = ITEM_PREFIX + VIEW_NAME;
         itemPrefix = ITEM_PREFIX;
         createMessage = STATION_CREATE;
         deleteMessage = STATION_DELETE;
+    }
+
+    private void initializeSelections() {
+        List<String> descriptions = new ArrayList<>(Arrays.asList(
+                MENU_CREATE, MENU_DELETE, MENU_READ, MENU_ESCAPE
+        ));
+        List<String> menuIndexs = new ArrayList<>(IntStream.range(MENU_START_INDEX, descriptions.size())
+                .mapToObj(Integer::toString)
+                .collect(Collectors.toList()));
+        menuIndexs.add(MENU_ESCAPE_VALUE);
+        initializeSelections(menuIndexs, descriptions);
     }
 
     public static StationView getInstance() {
